# samples/tokens/swagger.yaml
openapi: 3.0.3
info:
  title: Tokens
  version: "1.0"
  description: |-
    Issue, hold, transfer and redeem tokens. Backed by a distributed ledger,
    preserving privacy using Zero Knowledge Proofs. This API definition fronts 4 different services, which each play a different
    role in the ecosystem. 

    To try it out, let the Issuer issue some funds to one of the users. Watch the balances
    get updated and see the transaction history. Transfer it to another account, or redeem
    the tokens to destroy them.
servers:
  # - url: http://localhost:9500/
  #   description: auditor
  - url: http://localhost:9100/
    description: issuer
  - url: http://localhost:9500/
    description: alice and bob
  - url: http://localhost:9600/
    description: carlos and dan
paths:
...
  /endorser/init:
    servers:
      - url: http://localhost:9300/
        description: endorser 1
    post:
      tags:
        - endorser
      operationId: init
      summary: Initialize the network by committing the token parameters to the ledger.
      responses:
        "200":
          $ref: "#/components/responses/HealthSuccess"
        default:
          $ref: "#/components/responses/ErrorResponse"

  # # Auditor
  # /auditor/accounts/{id}:
  #   servers:
  #     - url: http://localhost:9000/
  #       description: auditor
  #   get:
  #     tags:
  #       - auditor
  #     parameters:
  #       - $ref: "#/components/parameters/id"
  #       - $ref: "#/components/parameters/code"
  #         required: true
  #     responses:
  #       "200":
  #         $ref: "#/components/responses/AccountSuccess"
  #       default:
  #         $ref: "#/components/responses/ErrorResponse"
  #     operationId: auditorAccount
  #     summary: Get an account and their balance of a certain type

  # /auditor/accounts/{id}/transactions:
  #   servers:
  #     - url: http://localhost:9000/
  #       description: auditor
  #   get:
  #     tags:
  #       - auditor
  #     parameters:
  #       - $ref: "#/components/parameters/id"
  #     operationId: auditorTransactions
  #     summary: Get all transactions for an account
  #     responses:
  #       "200":
  #         $ref: "#/components/responses/TransactionsSuccess"
  #       default:
  #         $ref: "#/components/responses/ErrorResponse"

  # Issuer
...
components:
  responses:
    AccountSuccess:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - payload
            properties:
              message:
                type: string
              payload:
                $ref: "#/components/schemas/Account"
...
    TransactionsSuccess:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - payload
            properties:
              message:
                type: string
                example: "transactions retrieved"
              payload:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionRecord"
...
    HealthSuccess:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                description: ok
            example:
              message: ok
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Amount:
      description: The amount to issue, transfer or redeem.
      type: object
      required:
        - code
        - value
      properties:
        code:
          description: the code of the token
          default: EURX
          type: string
        value:
          description: value in base units (usually cents)
          type: integer
          format: uint64
          minimum: 0
      example:
        code: EURX
        value: 100
    Account:
      type: object
      description: Information about an account and its balance
      required:
        - id
        - balance
      properties:
        id:
          type: string
          description: account id as registered at the Certificate Authority
          example: alice
        balance:
          type: array
          description: balance in base units for each currency
          items:
            $ref: "#/components/schemas/Amount"
      example:
        id: alice
        balance:
          - code: EURX
            value: 10000
          - code: USDX
            value: 200
...
    TransactionRecord:
      type: object
      description: A transaction
      required:
        - id
        - sender
        - recipient
        - amount
        - timestamp
        - status
        - message
      properties:
        id:
          type: string
          description: transaction id
        sender:
          type: string
          description: the sender of the transaction
        recipient:
          type: string
          description: the recipient of the transaction
        amount:
          $ref: "#/components/schemas/Amount"
        timestamp:
          type: string
          format: date-time
          description: 'timestamp in the format: "2018-03-20T09:12:28Z"'
        status:
          type: string
          description: Unknown | Pending | Confirmed | Deleted
        message:
          type: string
          description: user provided message
      example:
        id: 123
        sender: alice
        recipient: bob
        amount:
          code: EURX
          value: 100
        timestamp: "2018-03-20T09:12:28Z"
        status: Confirmed
        message: ""

    # Owner
...
    Error:
      required:
        - message
        - payload
      type: object
      properties:
        message:
          description: High level error message
          type: string
        payload:
          description: Details about the error
          type: string
      example:
        message: error message
        payload: ""
  parameters:
    id:
      name: id
      schema:
        example: alice
        description: account id as registered at the Certificate Authority
        type: string
      in: path
      required: true
    code:
      name: code
      in: query
      schema:
        example: EURX
        description: The token code to filter on
        type: string
  securitySchemes: {}
...
security: []